+++ capa/features/extractors/dnfile/insn.py
@@ -27,7 +27,14 @@ from dncil.cil.opcode import OpCodes
 
 import capa.features.extractors.helpers
 from capa.features.insn import API, Number, Property
-from capa.features.common import Class, String, Feature, Namespace, FeatureAccess, Characteristic
+from capa.features.common import (
+    Class,
+    String,
+    Feature,
+    Namespace,
+    FeatureAccess,
+    Characteristic,
+)
 from capa.features.address import Address
 from capa.features.extractors.dnfile.types import DnType, DnUnmanagedMethod
 from capa.features.extractors.base_extractor import BBHandle, InsnHandle, FunctionHandle
@@ -47,14 +54,19 @@ def get_callee(
     token_: int
     if token.table == dnfile.mdtable.MethodSpec.number:
         # map MethodSpec to MethodDef or MemberRef
-        row: Union[dnfile.base.MDTableRow, InvalidToken, str] = resolve_dotnet_token(pe, token)
-        assert isinstance(row, dnfile.mdtable.MethodSpecRow)
+        row: Union[dnfile.base.MDTableRow, InvalidToken, str] = resolve_dotnet_token(
+            pe, token
+        )
+        if not isinstance(row, dnfile.mdtable.MethodSpecRow):
+            return None
 
         if row.Method.table is None:
             logger.debug("MethodSpec[0x%X] Method table is None", token.rid)
             return None
 
-        token_ = calculate_dotnet_token_value(row.Method.table.number, row.Method.row_index)
+        token_ = calculate_dotnet_token_value(
+            row.Method.table.number, row.Method.row_index
+        )
     else:
         token_ = token.value
 
@@ -68,7 +80,9 @@ def get_callee(
     return callee
 
 
-def extract_insn_api_features(fh: FunctionHandle, bh, ih: InsnHandle) -> Iterator[tuple[Feature, Address]]:
+def extract_insn_api_features(
+    fh: FunctionHandle, bh, ih: InsnHandle
+) -> Iterator[tuple[Feature, Address]]:
     """parse instruction API features"""
     if ih.inner.opcode not in (
         OpCodes.Call,
@@ -78,7 +92,9 @@ def extract_insn_api_features(fh: FunctionHandle, bh, ih: InsnHandle) -> Iterato
     ):
         return
 
-    callee: Optional[Union[DnType, DnUnmanagedMethod]] = get_callee(fh.ctx["pe"], fh.ctx["cache"], ih.inner.operand)
+    callee: Optional[Union[DnType, DnUnmanagedMethod]] = get_callee(
+        fh.ctx["pe"], fh.ctx["cache"], ih.inner.operand
+    )
     if isinstance(callee, DnType):
         # ignore methods used to access properties
         if callee.access is None:
@@ -86,33 +102,48 @@ def extract_insn_api_features(fh: FunctionHandle, bh, ih: InsnHandle) -> Iterato
             yield API(str(callee)), ih.address
     elif isinstance(callee, DnUnmanagedMethod):
         # like kernel32.CreateFileA
-        for name in capa.features.extractors.helpers.generate_symbols(callee.module, callee.method):
+        for name in capa.features.extractors.helpers.generate_symbols(
+            callee.module, callee.method
+        ):
             yield API(name), ih.address
 
 
-def extract_insn_property_features(fh: FunctionHandle, bh, ih: InsnHandle) -> Iterator[tuple[Feature, Address]]:
+def extract_insn_property_features(
+    fh: FunctionHandle, bh, ih: InsnHandle
+) -> Iterator[tuple[Feature, Address]]:
     """parse instruction property features"""
     name: Optional[str] = None
     access: Optional[str] = None
 
     if ih.inner.opcode in (OpCodes.Call, OpCodes.Callvirt, OpCodes.Jmp):
         # property access via MethodDef or MemberRef
-        callee: Optional[Union[DnType, DnUnmanagedMethod]] = get_callee(fh.ctx["pe"], fh.ctx["cache"], ih.inner.operand)
+        callee: Optional[Union[DnType, DnUnmanagedMethod]] = get_callee(
+            fh.ctx["pe"], fh.ctx["cache"], ih.inner.operand
+        )
         if isinstance(callee, DnType):
             if callee.access is not None:
                 name = str(callee)
                 access = callee.access
 
-    elif ih.inner.opcode in (OpCodes.Ldfld, OpCodes.Ldflda, OpCodes.Ldsfld, OpCodes.Ldsflda):
+    elif ih.inner.opcode in (
+        OpCodes.Ldfld,
+        OpCodes.Ldflda,
+        OpCodes.Ldsfld,
+        OpCodes.Ldsflda,
+    ):
         # property read via Field
-        read_field: Optional[Union[DnType, DnUnmanagedMethod]] = fh.ctx["cache"].get_field(ih.inner.operand.value)
+        read_field: Optional[Union[DnType, DnUnmanagedMethod]] = fh.ctx[
+            "cache"
+        ].get_field(ih.inner.operand.value)
         if read_field is not None:
             name = str(read_field)
             access = FeatureAccess.READ
 
     elif ih.inner.opcode in (OpCodes.Stfld, OpCodes.Stsfld):
         # property write via Field
-        write_field: Optional[Union[DnType, DnUnmanagedMethod]] = fh.ctx["cache"].get_field(ih.inner.operand.value)
+        write_field: Optional[Union[DnType, DnUnmanagedMethod]] = fh.ctx[
+            "cache"
+        ].get_field(ih.inner.operand.value)
         if write_field is not None:
             name = str(write_field)
             access = FeatureAccess.WRITE
@@ -175,18 +206,25 @@ def extract_insn_namespace_class_features(
         type_ = fh.ctx["cache"].get_type(ih.inner.operand.value)
 
     if isinstance(type_, DnType):
-        yield Class(DnType.format_name(type_.class_, namespace=type_.namespace)), ih.address
+        yield (
+            Class(DnType.format_name(type_.class_, namespace=type_.namespace)),
+            ih.address,
+        )
         if type_.namespace:
             yield Namespace(type_.namespace), ih.address
 
 
-def extract_insn_number_features(fh, bh, ih: InsnHandle) -> Iterator[tuple[Feature, Address]]:
+def extract_insn_number_features(
+    fh, bh, ih: InsnHandle
+) -> Iterator[tuple[Feature, Address]]:
     """parse instruction number features"""
     if ih.inner.is_ldc():
         yield Number(ih.inner.get_ldc()), ih.address
 
 
-def extract_insn_string_features(fh: FunctionHandle, bh, ih: InsnHandle) -> Iterator[tuple[Feature, Address]]:
+def extract_insn_string_features(
+    fh: FunctionHandle, bh, ih: InsnHandle
+) -> Iterator[tuple[Feature, Address]]:
     """parse instruction string features"""
     if not ih.inner.is_ldstr():
         return
@@ -208,15 +246,21 @@ def extract_unmanaged_call_characteristic_features(
     if ih.inner.opcode not in (OpCodes.Call, OpCodes.Callvirt, OpCodes.Jmp):
         return
 
-    row: Union[str, InvalidToken, dnfile.base.MDTableRow] = resolve_dotnet_token(fh.ctx["pe"], ih.inner.operand)
+    row: Union[str, InvalidToken, dnfile.base.MDTableRow] = resolve_dotnet_token(
+        fh.ctx["pe"], ih.inner.operand
+    )
     if not isinstance(row, dnfile.mdtable.MethodDefRow):
         return
 
-    if any((row.Flags.mdPinvokeImpl, row.ImplFlags.miUnmanaged, row.ImplFlags.miNative)):
+    if any(
+        (row.Flags.mdPinvokeImpl, row.ImplFlags.miUnmanaged, row.ImplFlags.miNative)
+    ):
         yield Characteristic("unmanaged call"), ih.address
 
 
-def extract_features(fh: FunctionHandle, bbh: BBHandle, ih: InsnHandle) -> Iterator[tuple[Feature, Address]]:
+def extract_features(
+    fh: FunctionHandle, bbh: BBHandle, ih: InsnHandle
+) -> Iterator[tuple[Feature, Address]]:
     """extract instruction features"""
     for inst_handler in INSTRUCTION_HANDLERS:
         for feature, addr in inst_handler(fh, bbh, ih):